add arena tests
